+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/admin/drawback/grabfile/form/TGrabFileSqlForm.java
@@ -40,10 +40,10 @@ public class TGrabFileSqlForm extends DataTablesParamForm {
 	private String fileSize;
 
 	/**文件类型(1-文件夹;2-文件;)*/
-	private long type;
+	private int type;
 
 	/**状态(1-已删除;2-已启用;)*/
-	private long status;
+	private int status;
 
 	/**创建时间*/
 	private Date createTime;
@@ -61,7 +61,7 @@ public class TGrabFileSqlForm extends DataTablesParamForm {
 	private long sort;
 
 	/**散团类型*/
-	private long groupType;
+	private int groupType;
 
 	@Override
 	public Sql sql(SqlManager sqlManager) {
